assets: dynamic import customized components
closes #234
