+++ invenio_app_rdm/theme/assets/semantic-ui/js/invenio_app_rdm/search/ResultsGrid.item.js
@@ -1,0 +1,31 @@
+// This file is part of InvenioRDM
+// Copyright (C) 2020 CERN.
+// Copyright (C) 2020 Northwestern University.
+//
+// Invenio App RDM is free software; you can redistribute it and/or modify it
+// under the terms of the MIT License; see LICENSE file for more details.
+
+import React from "react";
+import { Card } from "semantic-ui-react";
+import _ from "lodash";
+import _truncate from "lodash/truncate";
+
+const RDMRecordResultsGridItem = ({ result, index }) => {
+  const description = _.get(
+    result,
+    "metadata.descriptions[0].description",
+    "No description"
+  );
+  return (
+    <Card fluid key={index} href={`/records/${result.id}`}>
+      <Card.Content>
+        <Card.Header>{result.metadata.titles[0].title}</Card.Header>
+        <Card.Description>
+          {_truncate(description, { length: 200 })}
+        </Card.Description>
+      </Card.Content>
+    </Card>
+  );
+};
+
+export default RDMRecordResultsGridItem;
