+++ invenio_app_rdm/theme/assets/semantic-ui/js/invenio_app_rdm/search/index.js
@@ -1,0 +1,16 @@
+// This file is part of InvenioRDM
+// Copyright (C) 2020 CERN.
+// Copyright (C) 2020 Northwestern University.
+//
+// Invenio App RDM is free software; you can redistribute it and/or modify it
+// under the terms of the MIT License; see LICENSE file for more details.
+
+import RDMRecordResultsListItem from "./ResultsList.item";
+import RDMRecordResultsGridItem from "./ResultsGrid.item";
+import RDMRecordSearchBarElement from "./SearchBar.element";
+
+export {
+  RDMRecordResultsListItem,
+  RDMRecordResultsGridItem,
+  RDMRecordSearchBarElement,
+};
